refactor(quick-260608-qe2): relocate TradingDirection and SystemStatus into core/enums/
- Add core/enums/trading.py (TradingDirection) and core/enums/system.py (SystemStatus)
- Re-export both from core/enums/__init__.py barrel
- sizing.py re-exports TradingDirection (6 call sites untouched); drop now-unused Enum import
- live_trading_system.py imports SystemStatus from core/enums; drop now-unused Enum import
